Move intercept cmd from cli to userd
