+++ pkg/client/cli/cmd_preview.go
@@ -66,7 +66,7 @@ func previewCommand() *cobra.Command {
 					if err != nil {
 						return err
 					}
-					fmt.Println(DescribeIntercepts([]*manager.InterceptInfo{intercept}, nil, false))
+					fmt.Println(cliutil.DescribeIntercepts([]*manager.InterceptInfo{intercept}, nil, false))
 					return nil
 				})
 			})
@@ -92,7 +92,7 @@ func previewCommand() *cobra.Command {
 					if err != nil {
 						return err
 					}
-					fmt.Println(DescribeIntercepts([]*manager.InterceptInfo{intercept}, nil, false))
+					fmt.Println(cliutil.DescribeIntercepts([]*manager.InterceptInfo{intercept}, nil, false))
 					return nil
 				})
 			})
